feat(#408): Jahresurlaubsanspruch mit Nachkommastelle (User.vacation_days dezimal)
Kundenreport (philvdb): eine 3-Tage-Teilzeitkraft hat 28×3/5 = 16,8 Urlaubstage.
Bisher war der Jahresanspruch nur ganzzahlig eingebbar → Rundung auf 17 → 0,2
Tage Ungerechtigkeit vs. einer Kollegin mit unterjährigem Eintritt.

- Migration 066: users.vacation_days Integer → Numeric(4,1) (up→down→up auf PG18
  verifiziert; Spalte = numeric(4,1); 16,8 speichert als 16,8).
- Schema: vacation_days int → float (UserBase/UserUpdate/UserListResponse,
  ge=0/le=50); suggested_vacation_days bleibt int (30×wd/5, immer ganzzahlig).
- Frontend UserForm: Eingabe parseInt → parseFloat + step="0.1".
- Der Calc rechnete bereits mit Decimal(str(user.vacation_days)) — unverändert.

DSGVO-Export-Fix (Folge der Numeric-Umstellung, Review-CRITICAL): zwei rohe
json.dumps-Export-Pfade schrieben vacation_days ungecastet als Decimal →
"Object of type Decimal is not JSON serializable" (HTTP 500 für JEDEN Nutzer,
auch bei ganzzahligem Wert):
- lifecycle_service._user_dict (Art.15-Selbstexport + Admin-Tenant-Export)
- auth.py /me/export (Art.20-Datenübertragbarkeit)
Beide casten jetzt float(...) — exakt wie das benachbarte weekly_hours (auch
Numeric). Vorher war vacation_days Integer, daher kein Cast nötig.

Verifiziert: Backend-Voll-Suite 1378 passed (die 6 verbleibenden Fails sind
pre-existing shift_planning-„MyToday"-Tests, date-abhängig, failen auch auf
master); Frontend tsc + UserForm 13/13 + build; PG18 Migration up→down→up +
numeric(4,1)-Round-Trip. Neuer Test tests/test_vacation_days_decimal.py.
Multi-Agent-Review: 1 CRITICAL (der Export-Decimal-Crash) gefunden + gefixt.

+++ backend/app/schemas/user.py
@@ -35,7 +35,7 @@ class UserBase(BaseModel):
     first_name: str = Field(..., min_length=1, max_length=100)
     last_name: str = Field(..., min_length=1, max_length=100)
     weekly_hours: float = Field(..., ge=0, le=60)
-    vacation_days: int = Field(..., ge=0, le=50)
+    vacation_days: float = Field(..., ge=0, le=50)  # #408: Dezimal (z. B. 16,8 für 3-Tage-Teilzeit)
     work_days_per_week: int = Field(default=5, ge=1, le=7)
     track_hours: bool = True
     calendar_color: str = Field(default='#93C5FD', pattern=r'^#[0-9A-Fa-f]{6}$')
@@ -103,7 +103,7 @@ class UserUpdate(BaseModel):
     first_name: Optional[str] = Field(None, min_length=1, max_length=100)
     last_name: Optional[str] = Field(None, min_length=1, max_length=100)
     weekly_hours: Optional[float] = Field(None, ge=0, le=60)
-    vacation_days: Optional[int] = Field(None, ge=0, le=50)
+    vacation_days: Optional[float] = Field(None, ge=0, le=50)  # #408: Dezimal
     work_days_per_week: Optional[int] = Field(None, ge=1, le=7)
     track_hours: Optional[bool] = None
     is_active: Optional[bool] = None
@@ -192,7 +192,7 @@ class UserListResponse(BaseModel):
     last_name: str
     email: Optional[str] = None
     weekly_hours: float
-    vacation_days: int
+    vacation_days: float  # #408: Dezimal
     work_days_per_week: int
     track_hours: bool
     is_active: bool
